Don't build duplicate prices in VatPriceGenerator
Spree::Variant::VatPriceGenerator does create duplicate prices when
called several times without the variant being saved in between.

This commit searches for existing prices in `Variant#prices` instead of
querying the database. That way, prices which are not yet persisted to
the database will be included in the search for existing prices.

Resolves: #1634